// src/app/shared/model/credit-review/crd-expsr-app-agr-hist-obj.model.ts
export class CrdExpsrAppAgrHistObj {

    CustExpsrAppAgrHistId: number;
    CustExpsrHId: number;
    AppNo: string;
    AgrmntNo: string;
    GoLiveDt: Date;
    OfferingName: string;
    AppStat: string;
    AppStep: string;
    ContractStat: string;
    OsPrincipal: number;
    Fpd: string;
    CustName: string;
    RoleCust: string;
    OverdueDays: number;
    OverdueAmt: number;
    constructor() {
    }
}